More precise configuration of resource constraints
The 'uvsnake' script sets default constraints, and
the long-running/memory intensive commands
(unoise3, uparse, otutab) were benchmarked and
now have more 'intelligent' constraints that change
depending on input file size. This is of course an approximation
that depends on the community complexity,
sequence quality, compression ratios and probably
other properties. But the constraints are quite
liberal and we suggest supplying -T/--retries on
computer clusters.